Moving frameworks to utils folder
The go files below are moved to the utils folder.

+ db.go
+ logging.go

Have created a file called constants, which are global , keeping it here unless someone finds a better solution on where to keep rather than importing the constants on around the place.

Done as part of the issue : https://github.com/pivotal-gss/gpmt2/issues/8